requirements.txt
albucore==0.0.13
albumentations==1.4.10
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.8.0
astor==0.8.1
beautifulsoup4==4.12.3
blis==1.2.0
catalogue==2.0.10
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
confection==0.1.5
contourpy==1.3.1
contractions==0.1.73
cycler==0.12.1
cymem==2.0.11
Cython==3.0.11
decorator==5.1.1
distro==1.9.0
easyocr==1.7.2
exceptiongroup==1.2.2
filelock==3.17.0
fire==0.7.0
fonttools==4.55.6
fsspec==2024.12.0
groq==0.18.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
imgaug==0.4.0
isodate==0.7.2
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
lmdb==1.6.2
lxml==5.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.12
networkx==3.4.2
ninja==1.11.1.3
nltk==3.9.1
numpy==2.2.2
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
packaging==24.2
paddle-bfloat==0.1.7
paddleocr==2.9.1
paddlepaddle==2.6.2
pillow==11.1.0
preshed==3.0.9
protobuf==5.29.3
pyahocorasick==2.1.0
pyclipper==1.3.0.post6
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
PyYAML==6.0.2
RapidFuzz==3.11.0
rdflib==7.1.3
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
rich==13.9.4
scikit-image==0.25.1
scikit-learn==1.6.1
scipy==1.15.1
shapely==2.0.6
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.38
srsly==2.5.1
sympy==1.13.1
termcolor==2.5.0
textsearch==0.0.24
thinc==8.3.4
threadpoolctl==3.5.0
tifffile==2025.1.10
tomli==2.2.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.17.2
